Make heap-fragmentation log 64-bit-safe (pointer hygiene, stage 3f)
CalculateHeapFragmentationText casts each allocation's pvBase to uint32
and page-masks it to build a histogram of touched pages. On x64 the bare
(uint32)pvBase truncates the pointer and warns. The page value is a
diagnostic key only (stored in a uint32[] and compared for equality to
count unique pages, never turned back into a pointer), so keep the 32-bit
key and just make the narrowing explicit with (uint32)(uintptr)pvBase.

No mask-width trap here (unlike FastHeap): the 0xfffff000 mask applies to
the already-narrowed uint32, not to a uintptr, so it stays correct.

The MemoryLog.hpp counter macros are untouched: their (long)/(int)/(
unsigned long) casts are on byte counts (mem, _msize, size), not pointers.

Behaviour-identical on Win32 where uintptr == uint32: built green (exit 0,
no new warnings), smoke-launched OK. Sixth slice of the uint32->uintptr
sweep after VirtualMem (ba41823), FastHeap (3825cd3), loaders (43c4b87),
audio (1d560a9), Mesh (0491ee3).

+++ jp2_pc/Source/Lib/Sys/MemoryLog.cpp
@@ -523,7 +523,7 @@ char* CalculateHeapFragmentationText()
 		u4_total += AllocList[i].nSize;
 		
 		int32 i4_size = AllocList[i].nSize;
-		uint32 u4_adr = ((uint32)AllocList[i].pvBase) & 0xfffff000;
+		uint32 u4_adr = ((uint32)(uintptr)AllocList[i].pvBase) & 0xfffff000;
 		while (i4_size>0)
 		{
 			au4PageBase[u4_pages] = u4_adr;
